requirements.txt
pandas
opencv-python
torch
torchvision
torchsummary
pillow
colorama